Refactor usage of history components to integrate CredentialHistory
